+++ guidance/library/__init__.py
@@ -15,7 +15,7 @@ from .._grammar import byte_range
 
 # context blocks
 from ._block import block
-from ._role import role, system, assistant, user, function, instruction, indent_roles
+from ._role import role, system, assistant, user #, function, instruction, indent_roles
 from ._format import monospace
 from ._silent import silent
 from ._set_var import set_var
