+++ guidance/library/_role.py
@@ -1,34 +1,8 @@
-from .._guidance import guidance
-from ._block import block
-from ._set_attribute import set_attribute
+from contextlib import AbstractContextManager
+from ..models.base import role
 
 
-@guidance
-def role_opener(lm, role_name, **kwargs):
-    lm += lm.role_opener(role_name, **kwargs)
-    return lm
-
-
-@guidance
-def role_closer(lm, role_name, **kwargs):
-    lm += lm.role_closer(role_name, **kwargs)
-    return lm
-
-
-# TODO HN: Add a docstring to better describe arbitrary role functions
-def role(role_name, text=None, **kwargs):
-    if text is None:
-        return block(
-            name=role_name,
-            opener=role_opener(role_name, **kwargs),
-            closer=role_closer(role_name, **kwargs),
-        )
-    else:
-        assert False
-        # return self.append(open_text + text + close_text)
-
-
-def system(text=None, **kwargs):
+def system() -> AbstractContextManager[None]:
     """Indicate the 'system' prompt
 
     A convention has grown up around 'chat' APIs that
@@ -41,10 +15,10 @@ def system(text=None, **kwargs):
         >>>     lm += "A system prompt"
 
     """
-    return role("system", text, **kwargs)
+    return role("system")
 
 
-def user(text=None, **kwargs):
+def user() -> AbstractContextManager[None]:
     """Indicate the 'user' prompt
 
     A convention has grown up around 'chat' APIs that
@@ -57,10 +31,10 @@ def user(text=None, **kwargs):
         >>>     lm += "What the user said"
 
     """
-    return role("user", text, **kwargs)
+    return role("user")
 
 
-def assistant(text=None, **kwargs):
+def assistant() -> AbstractContextManager[None]:
     """Indicate the 'assistant' prompt
 
     A convention has grown up around 'chat' APIs that
@@ -74,16 +48,4 @@ def assistant(text=None, **kwargs):
         >>>     lm += gen(name="model_output", max_tokens=20)
 
     """
-    return role("assistant", text, **kwargs)
-
-
-def function(text=None, **kwargs):
-    return role("function", text, **kwargs)
-
-
-def instruction(text=None, **kwargs):
-    return role("instruction", text, **kwargs)
-
-
-def indent_roles(indent=True):
-    return set_attribute("indent_roles", indent)
+    return role("assistant")
