+++ guidance/models/_azureai_studio.py
@@ -6,7 +6,7 @@ import diskcache as dc
 import platformdirs
 import requests
 
-from ._model import Chat
+from .base._engine import Chat
 from ._grammarless import GrammarlessEngine, Grammarless
