+++ guidance/models/transformers/__init__.py
@@ -1,1 +1,8 @@
-from ._transformers import Transformers, TransformersTokenizer
+from ._engine import TransformersEngine, TransformersTokenizer
+from ._model import Transformers
+
+__all__ = [
+    "Transformers",
+    "TransformersEngine",
+    "TransformersTokenizer",
+]
