+++ tests/model_integration/library/test_commit_point.py
@@ -1,8 +1,8 @@
 import pytest
-from guidance import Tool, capture, commit_point, models, select, string
+from guidance.models import Tool, capture, commit_point, Model, select, string
 
 @pytest.mark.xfail(reason="Commit points are not supported")
-def test_commit_point(selected_model: models.Model):
+def test_commit_point(selected_model: Model.Model):
     lm = selected_model
     tools = [Tool(callable=lambda x: x)]
     stop_pattern = select([string(lm.engine.tokenizer.eos_token)])
