+++ tests/model_integration/test_grammar.py
@@ -1,7 +1,7 @@
-from guidance import models, select
+from guidance.models import Model, select
 
 
-def test_select_simple(selected_model: models.Model):
+def test_select_simple(selected_model: Model.Model):
     lm = selected_model
     options = ["baad I think", "bad I think", "bad"]
     lm = lm + "Scott is quite " + select(name="bad", options=options)
