+++ tests/model_specific/test_transformers.py
@@ -1,7 +1,6 @@
 import pytest
 
 from guidance import gen, select, models, assistant, system, user, guidance
-
 from ..utils import get_model
